Reports import and run
